make versioning less confusing
