+++ structure_tests/structure_test_v0.go
@@ -15,21 +15,25 @@
 package main
 
 import (
+	"bytes"
 	"fmt"
 	"io/ioutil"
 	"os"
+	"os/exec"
+	"strings"
+	"syscall"
 	"testing"
 )
 
-type StructureTestv100 struct {
+type StructureTestv0 struct {
 	GlobalEnvVars      []EnvVar
-	CommandTests       []CommandTestv1
+	CommandTests       []CommandTestv0
 	FileExistenceTests []FileExistenceTestv0
 	FileContentTests   []FileContentTestv0
 	LicenseTests       []LicenseTestv0
 }
 
-func (st StructureTestv100) RunAll(t *testing.T) int {
+func (st StructureTestv0) RunAll(t *testing.T) int {
 	originalVars := SetEnvVars(t, st.GlobalEnvVars)
 	defer ResetEnvVars(t, originalVars)
 	testsRun := 0
@@ -40,19 +44,17 @@ func (st StructureTestv100) RunAll(t *testing.T) int {
 	return testsRun
 }
 
-func (st StructureTestv100) RunCommandTests(t *testing.T) int {
+func (st StructureTestv0) RunCommandTests(t *testing.T) int {
 	counter := 0
 	for _, tt := range st.CommandTests {
 		t.Run(tt.LogName(), func(t *testing.T) {
-			validateCommandTestv1(t, tt)
+			validateCommandTestv0(t, tt)
 			for _, setup := range tt.Setup {
 				ProcessCommand(t, tt.EnvVars, setup, tt.ShellMode, false)
 			}
 
-			fullCommand := append([]string{tt.Entrypoint}, tt.Args...)
-
-			stdout, stderr, exitcode := ProcessCommand(t, tt.EnvVars, fullCommand, tt.ShellMode, true)
-			CheckOutputv1(t, tt, stdout, stderr, exitcode)
+			stdout, stderr, exitcode := ProcessCommand(t, tt.EnvVars, tt.Command, tt.ShellMode, true)
+			CheckOutputv0(t, tt, stdout, stderr, exitcode)
 
 			for _, teardown := range tt.Teardown {
 				ProcessCommand(t, tt.EnvVars, teardown, tt.ShellMode, false)
@@ -63,7 +65,7 @@ func (st StructureTestv100) RunCommandTests(t *testing.T) int {
 	return counter
 }
 
-func (st StructureTestv100) RunFileExistenceTests(t *testing.T) int {
+func (st StructureTestv0) RunFileExistenceTests(t *testing.T) int {
 	counter := 0
 	for _, tt := range st.FileExistenceTests {
 		t.Run(tt.LogName(), func(t *testing.T) {
@@ -100,7 +102,7 @@ func (st StructureTestv100) RunFileExistenceTests(t *testing.T) int {
 	return counter
 }
 
-func (st StructureTestv100) RunFileContentTests(t *testing.T) int {
+func (st StructureTestv0) RunFileContentTests(t *testing.T) int {
 	counter := 0
 	for _, tt := range st.FileContentTests {
 		t.Run(tt.LogName(), func(t *testing.T) {
@@ -127,7 +129,7 @@ func (st StructureTestv100) RunFileContentTests(t *testing.T) int {
 	return counter
 }
 
-func (st StructureTestv100) RunLicenseTests(t *testing.T) int {
+func (st StructureTestv0) RunLicenseTests(t *testing.T) int {
 	for num, tt := range st.LicenseTests {
 		t.Run(tt.LogName(num), func(t *testing.T) {
 			checkLicenses(t, tt)
@@ -137,7 +139,78 @@ func (st StructureTestv100) RunLicenseTests(t *testing.T) int {
 	return 0
 }
 
-func CheckOutputv1(t *testing.T, tt CommandTestv1, stdout string, stderr string, exitCode int) {
+// given an array of command parts, construct a full command and execute it against the
+// current environment. a list of environment variables can be passed to be set in the
+// environment before the command is executed. additionally, a boolean flag is passed
+// to specify whether or not we care about the output of the command.
+func ProcessCommandv0(t *testing.T, envVars []EnvVar, fullCommand []string,
+	shellMode bool, checkOutput bool) (string, string, int) {
+	var cmd *exec.Cmd
+	if len(fullCommand) == 0 {
+		t.Logf("empty command provided: skipping...")
+		return "", "", -1
+	}
+	var command string
+	var flags []string
+	if shellMode {
+		command = "/bin/sh"
+		flags = []string{"-c", strings.Join(fullCommand, " ")}
+	} else {
+		command = fullCommand[0]
+		flags = fullCommand[1:]
+	}
+	originalVars := SetEnvVars(t, envVars)
+	defer ResetEnvVars(t, originalVars)
+	if len(flags) > 0 {
+		cmd = exec.Command(command, flags...)
+	} else {
+		cmd = exec.Command(command)
+	}
+
+	if checkOutput {
+		t.Logf("Executing: %s", cmd.Args)
+	} else {
+		t.Logf("Executing setup/teardown: %s", cmd.Args)
+	}
+
+	var outbuf, errbuf bytes.Buffer
+
+	cmd.Stdout = &outbuf
+	cmd.Stderr = &errbuf
+
+	err := cmd.Run()
+	stdout := outbuf.String()
+	if stdout != "" {
+		t.Logf("stdout: %s", stdout)
+	}
+	stderr := errbuf.String()
+	if stderr != "" {
+		t.Logf("stderr: %s", stderr)
+	}
+	var exitCode int
+	if err != nil {
+		if checkOutput {
+			// The test might be designed to run a command that exits with an error.
+			t.Logf("Error running command: %s. Continuing.", err)
+		} else {
+			t.Fatalf("Error running setup/teardown command: %s.", err)
+		}
+		switch err := err.(type) {
+		default:
+			t.Errorf("Command failed to start! Unable to retrieve error info!")
+		case *exec.ExitError:
+			exitCode = err.Sys().(syscall.WaitStatus).ExitStatus()
+		case *exec.Error:
+			// Command started but failed to finish, so we can at least check the stderr
+			stderr = err.Error()
+		}
+	} else {
+		exitCode = cmd.ProcessState.Sys().(syscall.WaitStatus).ExitStatus()
+	}
+	return stdout, stderr, exitCode
+}
+
+func CheckOutputv0(t *testing.T, tt CommandTestv0, stdout string, stderr string, exitCode int) {
 	for _, errStr := range tt.ExpectedError {
 		errMsg := fmt.Sprintf("Expected string '%s' not found in error!", errStr)
 		compileAndRunRegex(errStr, stderr, t, errMsg, true)
