+++ structure_tests/structure_test_v1.go
@@ -15,25 +15,21 @@
 package main
 
 import (
-	"bytes"
 	"fmt"
 	"io/ioutil"
 	"os"
-	"os/exec"
-	"strings"
-	"syscall"
 	"testing"
 )
 
-type StructureTestv000 struct {
+type StructureTestv1 struct {
 	GlobalEnvVars      []EnvVar
-	CommandTests       []CommandTestv0
+	CommandTests       []CommandTestv1
 	FileExistenceTests []FileExistenceTestv0
 	FileContentTests   []FileContentTestv0
 	LicenseTests       []LicenseTestv0
 }
 
-func (st StructureTestv000) RunAll(t *testing.T) int {
+func (st StructureTestv1) RunAll(t *testing.T) int {
 	originalVars := SetEnvVars(t, st.GlobalEnvVars)
 	defer ResetEnvVars(t, originalVars)
 	testsRun := 0
@@ -44,17 +40,19 @@ func (st StructureTestv000) RunAll(t *testing.T) int {
 	return testsRun
 }
 
-func (st StructureTestv000) RunCommandTests(t *testing.T) int {
+func (st StructureTestv1) RunCommandTests(t *testing.T) int {
 	counter := 0
 	for _, tt := range st.CommandTests {
 		t.Run(tt.LogName(), func(t *testing.T) {
-			validateCommandTestv0(t, tt)
+			validateCommandTestv1(t, tt)
 			for _, setup := range tt.Setup {
 				ProcessCommand(t, tt.EnvVars, setup, tt.ShellMode, false)
 			}
 
-			stdout, stderr, exitcode := ProcessCommand(t, tt.EnvVars, tt.Command, tt.ShellMode, true)
-			CheckOutputv0(t, tt, stdout, stderr, exitcode)
+			fullCommand := append([]string{tt.Entrypoint}, tt.Args...)
+
+			stdout, stderr, exitcode := ProcessCommand(t, tt.EnvVars, fullCommand, tt.ShellMode, true)
+			CheckOutputv1(t, tt, stdout, stderr, exitcode)
 
 			for _, teardown := range tt.Teardown {
 				ProcessCommand(t, tt.EnvVars, teardown, tt.ShellMode, false)
@@ -65,7 +63,7 @@ func (st StructureTestv000) RunCommandTests(t *testing.T) int {
 	return counter
 }
 
-func (st StructureTestv000) RunFileExistenceTests(t *testing.T) int {
+func (st StructureTestv1) RunFileExistenceTests(t *testing.T) int {
 	counter := 0
 	for _, tt := range st.FileExistenceTests {
 		t.Run(tt.LogName(), func(t *testing.T) {
@@ -102,7 +100,7 @@ func (st StructureTestv000) RunFileExistenceTests(t *testing.T) int {
 	return counter
 }
 
-func (st StructureTestv000) RunFileContentTests(t *testing.T) int {
+func (st StructureTestv1) RunFileContentTests(t *testing.T) int {
 	counter := 0
 	for _, tt := range st.FileContentTests {
 		t.Run(tt.LogName(), func(t *testing.T) {
@@ -129,7 +127,7 @@ func (st StructureTestv000) RunFileContentTests(t *testing.T) int {
 	return counter
 }
 
-func (st StructureTestv000) RunLicenseTests(t *testing.T) int {
+func (st StructureTestv1) RunLicenseTests(t *testing.T) int {
 	for num, tt := range st.LicenseTests {
 		t.Run(tt.LogName(num), func(t *testing.T) {
 			checkLicenses(t, tt)
@@ -139,78 +137,7 @@ func (st StructureTestv000) RunLicenseTests(t *testing.T) int {
 	return 0
 }
 
-// given an array of command parts, construct a full command and execute it against the
-// current environment. a list of environment variables can be passed to be set in the
-// environment before the command is executed. additionally, a boolean flag is passed
-// to specify whether or not we care about the output of the command.
-func ProcessCommandv0(t *testing.T, envVars []EnvVar, fullCommand []string,
-	shellMode bool, checkOutput bool) (string, string, int) {
-	var cmd *exec.Cmd
-	if len(fullCommand) == 0 {
-		t.Logf("empty command provided: skipping...")
-		return "", "", -1
-	}
-	var command string
-	var flags []string
-	if shellMode {
-		command = "/bin/sh"
-		flags = []string{"-c", strings.Join(fullCommand, " ")}
-	} else {
-		command = fullCommand[0]
-		flags = fullCommand[1:]
-	}
-	originalVars := SetEnvVars(t, envVars)
-	defer ResetEnvVars(t, originalVars)
-	if len(flags) > 0 {
-		cmd = exec.Command(command, flags...)
-	} else {
-		cmd = exec.Command(command)
-	}
-
-	if checkOutput {
-		t.Logf("Executing: %s", cmd.Args)
-	} else {
-		t.Logf("Executing setup/teardown: %s", cmd.Args)
-	}
-
-	var outbuf, errbuf bytes.Buffer
-
-	cmd.Stdout = &outbuf
-	cmd.Stderr = &errbuf
-
-	err := cmd.Run()
-	stdout := outbuf.String()
-	if stdout != "" {
-		t.Logf("stdout: %s", stdout)
-	}
-	stderr := errbuf.String()
-	if stderr != "" {
-		t.Logf("stderr: %s", stderr)
-	}
-	var exitCode int
-	if err != nil {
-		if checkOutput {
-			// The test might be designed to run a command that exits with an error.
-			t.Logf("Error running command: %s. Continuing.", err)
-		} else {
-			t.Fatalf("Error running setup/teardown command: %s.", err)
-		}
-		switch err := err.(type) {
-		default:
-			t.Errorf("Command failed to start! Unable to retrieve error info!")
-		case *exec.ExitError:
-			exitCode = err.Sys().(syscall.WaitStatus).ExitStatus()
-		case *exec.Error:
-			// Command started but failed to finish, so we can at least check the stderr
-			stderr = err.Error()
-		}
-	} else {
-		exitCode = cmd.ProcessState.Sys().(syscall.WaitStatus).ExitStatus()
-	}
-	return stdout, stderr, exitCode
-}
-
-func CheckOutputv0(t *testing.T, tt CommandTestv0, stdout string, stderr string, exitCode int) {
+func CheckOutputv1(t *testing.T, tt CommandTestv1, stdout string, stderr string, exitCode int) {
 	for _, errStr := range tt.ExpectedError {
 		errMsg := fmt.Sprintf("Expected string '%s' not found in error!", errStr)
 		compileAndRunRegex(errStr, stderr, t, errMsg, true)
